fixed modular encoder to properly propagate set_train; refactoring: merge model.py and decorators.py into hier_model.py
